move Config to separate file
